(#152) Added Level Manager class
