+++ data/Level/__init__.py
@@ -1,0 +1,1 @@
+from .LevelManager import LevelManager
